Server/requirements.txt
# LDPlayer Management System - Server Dependencies
# Версия: 1.0.0

# Core FastAPI dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.10.6
pydantic-settings==2.1.0

# Security & Authentication
PyJWT==2.10.1
cryptography==41.0.7
python-dotenv==1.0.0

# Remote management
pywinrm==0.5.0
paramiko==4.0.0
smbprotocol==1.10.1

# System monitoring & Resilience
psutil==5.9.6
prometheus-client==0.19.0
tenacity==8.2.3
pybreaker==1.4.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Database (optional, for future use)
sqlalchemy==2.0.23
alembic==1.13.1

# HTTP client
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Date and time utilities
python-dateutil==2.8.2

# Data validation
email-validator==2.1.0

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Authentication (PyJWT используется в коде, python-jose удален как дубликат)
passlib[bcrypt]==1.7.4

# Task queue (for future batch operations)
celery==5.3.4
redis==5.0.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
